Update dependencies
One should especially focus on the Sphinx update (from v5 to v7), rdflib
update (from v6 to v7) and the necessity to keep pydata-sphinx-theme at
its current version due to too many differences in the look and feel of
the docs if updated to the latest version.

Also, all currently unversioned dependencies have been fixed to a
version.